refactor(vss_core): keep memory adapters contract-only
Push SearchAdapter into search_core and leave summary/alert mappers
outside memory so the library owns protocol/helpers only.
